добавляет тест data-reducer
