+++ src/services/reducers/root-reducer.ts
@@ -1,6 +1,6 @@
 import { dataReducer } from './data-reducer/data-reducer';
 import constructorReducer from './constructor-reducer';
-import modalIngredientReducer from './modal-ingredient-reducer';
+import { modalIngredientReducer } from './modal-ingredient-reducer/modal-ingredient-reducer';
 import { orderReducer } from './order-reducer/order-reducer';
 import userReducer from './login-reducer';
 import orderSocketReducer from './order-socket-reducer';
